requirements.txt
torch
pytorch-lightning
python-telegram-bot
tensorboard